Remove unnecessary changes, move gptq detection function to 4bit.py
