+++ templates/popups.html
@@ -75,10 +75,6 @@
 				<input type="checkbox" data-toggle="toggle" data-onstyle="success" id="use_8_bit" checked>
 				<div class="box-label">Use 8 bit mode</div>
 			</div>
-			<div class="box flex-push-right hidden" id=use_4_bit_div>
-				<input type="checkbox" data-toggle="toggle" data-onstyle="success" id="use_4_bit" checked>
-				<div class="box-label">Use 4 bit mode</div>
-			</div>
 			<button type="button" class="btn popup_load_cancel_button action_button disabled" onclick="load_model()" id="btn_loadmodelaccept" disabled>Load</button>
 			<button type="button" class="btn popup_load_cancel_button" onclick='closePopups();' id="btn_loadmodelclose">Cancel</button>
 		</div>
